refactor(agent): Move SSH server into agentssh package
Refs: #6177
